Add the scaffolding for release promotion and a promote phase
The scaffolding part is heavily inspired by the code from
firefox-android and nothing's special in there.

The promote workflow is basically `SPM_Deploy_Prod_Beta` with the
shipping to appstoreconnect part ripped out. It runs on bitrise through
a [small taskcluster <=> bitrise
interface](https://github.com/mozilla-releng/scriptworker-scripts/tree/master/bitrisescript)

The scheme decision ended up being messy but the idea is as follows:

- If the repo URL contains staging, override the wanted scheme to
  `FirefoxStaging` and export method to `development`
- Otherwise, assume it's a production build, use the `BITRISE_SCHEME`
  passed in the environment for the task and change `CODE_SIGN_IDENTITY`
  to distribution.

The `BITRISE_SCHEME` is always set to either `Firefox` or `FirefoxBeta`
by the promote task itself (but is ignored on staging builds, as per the
above). It is set to `FirefoxBeta` if the build is specifically for a
beta version and `Firefox` otherwise. A `beta` version is any version
that has a patch component in it, `139.0.0` is a beta version, `139.0`
isn't. The source of truth for this is the version specified in the
`version.txt` for the promoted commit.

Because taskgraph doesn't support `by-` keys in lists, I had to write
a transform that would replace a well known value
(`<release-type-scheme>`) in every specified environment variable for
the promotion tasks.

The promotion task will output 2 artifacts:

- A `Client.ipa` containing the app to upload to TestFlight
- A `symbols.zip` archive containing the associated debug symbols to
  upload to sentry

Those artifacts are uploaded to bitrise and reexported by the
taskcluster task. The second part is important for the next phase.