Add a push phase to release promotion
The push phase it fairly straightforwad. It grabs the artifacts from the
promote phase and pushes them where it needs to.

That is sentry for the debug symbols and TestFlight for the app itself.
We get those artifacts from the promote task through taskcluster as
they're already there and we can easily reference it back.