Add FXIOS-13067 - shipit support for focus/klar
This is mostly straightforward, we kept the same idea for bitrise
workflows as we had for firefox. The main difference is that instead of
having a release type (beta/release), we have a product variant
(focus/klar). The shape of the graph is the same as for prod releases
where it would trigger beta/release promote/push at the same time but
instead it triggers focus/klar but we didn't want to overload the
release-type term by adding klar/focus to them.

Fixes #28463